backend/requirements.txt
fastapi
uvicorn
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-dotenv
